requirements.txt
asgiref==3.5.2
certifi==2022.5.18.1
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
Django==4.0.4
django-cors-headers==3.10.0
django-environ==0.8.1
django-filter==21.1
django-phonenumber-field==6.1.0
django-phonenumbers==1.0.1
django-rest-swagger==2.2.0
django-restframework==0.0.1
djangorestframework==3.13.1
drf-yasg==1.20.0
drfpasswordless==1.5.8
idna==3.3
inflection==0.5.1
install==1.3.5
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
openapi-codec==1.3.2
packaging==21.3
phonenumbers==8.12.48
Pillow==9.1.1
PyJWT==2.4.0
pyotp==2.6.0
pyparsing==3.0.9
pytz==2022.1
razorpay==1.3.0
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
sqlparse==0.4.2
twilio==7.9.1
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
